refactor: [#1182] move extracted service to new mod
